refactor(cookbooks): share PDF pipeline sample helpers
