+++ sdks/python/examples/pdf_pipeline/sample_pdf.py
@@ -1,0 +1,41 @@
+import io
+
+
+def make_sample_pdf(text: str) -> bytes:
+    """Create a minimal single-page PDF containing the given text."""
+    try:
+        from pypdf import PdfWriter
+        from pypdf.generic import (
+            DecodedStreamObject,
+            DictionaryObject,
+            NameObject,
+        )
+    except ImportError:
+        raise ImportError(
+            "pypdf is required for this example. "
+            "Install it in your Python environment before running."
+        )
+
+    writer = PdfWriter()
+    writer.add_blank_page(width=612, height=792)
+    page = writer.pages[0]
+
+    font = DictionaryObject()
+    font[NameObject("/Type")] = NameObject("/Font")
+    font[NameObject("/Subtype")] = NameObject("/Type1")
+    font[NameObject("/BaseFont")] = NameObject("/Helvetica")
+    font_ref = writer._add_object(font)
+
+    resources = page.get("/Resources", DictionaryObject())
+    if "/Font" not in resources:
+        resources[NameObject("/Font")] = DictionaryObject()
+    resources["/Font"][NameObject("/F1")] = font_ref
+    page[NameObject("/Resources")] = resources
+
+    stream = DecodedStreamObject()
+    stream.set_data(f"BT /F1 12 Tf 72 720 Td ({text}) Tj ET".encode())
+    page[NameObject("/Contents")] = writer._add_object(stream)
+
+    buf = io.BytesIO()
+    writer.write(buf)
+    return buf.getvalue()
